using integer math and added 1px of padding in the hopes that issue #414's new concern about underscores being clipped is fixed.

+++ src/widgets/QDisassemblyView.h
@@ -110,16 +110,16 @@ private:
 	bool get_line_of_address(edb::address_t addr, unsigned int& line) const;
 
 private:
-	std::shared_ptr<IRegion>                  region_;
+	std::shared_ptr<IRegion>          region_;
 	QVector<edb::address_t>           show_addresses_;
 	SyntaxHighlighter *const          highlighter_;
 	edb::address_t                    address_offset_;
 	edb::address_t                    selected_instruction_address_;
 	edb::address_t                    current_address_;
-	qreal                             font_height_; // height of a character in this font
-	qreal                             font_width_;  // width of a character in this font
-	qreal                             icon_width_;
-	qreal                             icon_height_;
+	int                               font_height_; // height of a character in this font
+	int                               font_width_;  // width of a character in this font
+	int                               icon_width_;
+	int                               icon_height_;
 	int                               line1_;
 	int                               line2_;
 	int                               line3_;
